fix pidigits1
most obscure JIT bug ever, as it was written before there was an
implicit promote of y.

reason: after inlining the old code looked like this:

if y == 0:
    ...
elif y == -1: # overflow check in int_floordiv_ovf
    ...
else:
    ...

this would be turned into a switch on y, with three cases. Switches
promote the switched upon variable. Obviously this is nonsense in this
case, so this is a bug that needs fixing in RPython. For now, work
around it.