start implementing denormalized rationals:
whenever we compute (/ a b) where a and be are fixnums, we don't produce
a normal rational, because that involves constructing bignums and
computing a GCD. Instead, we produce a "denormalized rational", which
just stores a and b. All operations on that will turn it into a regular
rational, except floor, which just returns a // b.