disable caching for the linklet vars from jsons
we're keeping the rpython regexp functions for the old pycket, and
they are exposed to the prim_env before anything else happens. As a
result, the linklet vars in the expander linklet for functions like
"regexp-match" bound the rpython procedures (at the expander
instantiation), and because of the caching in the variable they don't
get the new procedures (coming from the regexp linklet). This break
things because a "regexp" in the regexp linklet is a struct
(rx:regexp), not a value per se (e.g. W_Regexp), so the regexp
functions (like regexp-match) defined in RPython and Racket are not
interchangable.

That's why we disable the caching on the linklet vars that are
generated by the expand.py, while producing ast from
json. Consequently those linklet vars never cache their values, which
is a performance issue. However, that is going to be resolved by #229

For reasons that are unclear to me at the moment, we also need to
instantiate the regexp linklet before the expander linklet for this to
work. (Maybe ModuleVars are involved too, which would make sense
because they're always caching)