avoid having regexp literals in the generated json
because a regexp in the regexp linklet is a struct (rx:regexp)(not a
primitive value), for each regexp literal we produce an application
AST for the "regexp" function (and friends) to produce a rx:regexp in
the runtime.

Note that, the regexp linklet also contains regexp literals, so
synthesizing a rx:regexp from a regexp literal in the expand.py
wouldn't work (since we don't have the rx:regexp struct by then).