having separate implementations for partially simple primitives for stackful (for them to not ConvertStack) might be the cleanest solution ...
... We can't use the current method dispatch based implementation (for
instance for vector-ref) because 1) it's hacky to make all the
env/cont parameters optional everywhere (plus the extra_info's
position is problem) but even if we do that 2) the return type of all
the methods is going to be CEK triplets, but the calls coming from the
stackful can't return that without env & cont. Besides, the app callers
(call_with_extra_info_and_stack) all expect and return a single value.