enter the stackful from begin as well
there are 2 ways from exiting the stackful, 1) ConvertStack, 2) normal
return. If we don't enter the stackful from Begin as well, then
normally returning calls get back to the BeginCont created at CEK and
continue from there (eventually getting back into the stackful
again). This change allows it to stay in the stackful as long as we're
inside the Begin (which is usually true when we're inside a 'for' loop
at the Racket level)