trying to make new-simple to be called in CPS...
...demonstrates that wherever we place a dynamic decision on calling a
primitive in two different conventions we will have a "return type"
problem, where the non-simple version is going to return a tuple and
the simple version is going to return a value. Therefore, we either
need to unify the results (e.g. turn the CEK tuple into a Racket
value) or have different exposed primitives (having their own callers
-like 'simple_func'- with a corresponding return type) that will
eventually be called from the correct path implementing the
appropriate convention.