feat(linklet): implement linklet variables with W_Cell
W_LinkletVar is no longer used for linklet variables. This is based
on an observation that linklet variables are just like cells.

We currently put W_Cell(W_LinkletVar(value)) into the toplevel
environment for a linklet variable. With this change, we just put
W_Cell(value), which not only saves a bunch of operations in traces,
but also allows every linklet variable to benefit from the strategies
implemented on W_Cells.